Separate the tests into multiple files.
